Fixed app crash caused by the SoundInput class and other fixes
- Fixed app crash caused by the SoundInput class
- Saves have the ability to not replace some variables in PartyControlSimple

+++ LightParty/Party/PartyUIUpdater.cs
@@ -163,7 +163,7 @@ namespace LightParty.Party
                 useIntervalPageTwo = true;
             }
 
-            randomInterval = PartyOptions.activePartyOption.randomInterval;
+            randomInterval = (float)PartyOptions.activePartyOption.randomInterval;
             UpdateRandomIntervalTextBoxes();
         }
